2.x: Common test classes for Processors.
